Iteration 1 Complete + Note on repository change
Sprint 1's functionality is complete and project notebook updates are
now found in CS449-ProjectNotebook-AlAshby.doc

Due to a mishap while attempting to remove a commit that disallowed me
to sync the repository, I had to create a new one. The previous commits
can still be viewed at http://www.github.com/alashby/CS449-Project.
Hopefully this will  be the only instance of a change in repository,